Updates onBeforeMount to within the Report component.
