+++ index.html
@@ -10,7 +10,10 @@
       property="og:description"
       content="View historical sea ice data from the seas around the circumpolar north and discover how ice extent and concentration have changed over time."
     />
-    <meta property="og:image" content="https://snap.uaf.edu/tools/sea-ice-atlas/hsia-og-image.png" />
+    <meta
+      property="og:image"
+      content="https://snap.uaf.edu/tools/sea-ice-atlas/hsia-og-image.png"
+    />
     <link rel="icon" href="/favicon.ico" />
     <!-- TODO this stylesheet should be imported as a module dependency -->
     <link
@@ -35,9 +38,8 @@
         >We&rsquo;re sorry but this site doesn&rsquo;t work properly without JavaScript enabled.
         Please enable it to continue. If you have questions or want help accessing this content
         through different media, please email us at
-        <a href="mailto:uaf-snap-data-tools@alaska.edu">uaf-snap-data-tools@alaska.edu</a>.</strong>
-    </p>
-  </noscript>
-</body>
-
+        <a href="mailto:uaf-snap-data-tools@alaska.edu">uaf-snap-data-tools@alaska.edu</a>.</strong
+      >
+    </noscript>
+  </body>
 </html>
