+++ src/views/ReportView.vue
@@ -5,7 +5,7 @@
       <ReportTitle v-bind:class="{ 'is-hidden': !validMapPixel }" />
       <LoadingBlock />
       <InvalidPlace />
-      <Report v-bind:class="{ 'is-hidden': !validMapPixel }" />
+      <Report v-bind:class="{ 'is-hidden': !validMapPixel }" :lat="props.lat" :lng="props.lng" />
     </div>
   </section>
 </template>
@@ -18,18 +18,8 @@ import Report from '../components/Report.vue'
 import BackButton from '../components/BackButton.vue'
 import { useAtlasStore } from '@/stores/atlas'
 import { storeToRefs } from 'pinia'
-import { onBeforeMount } from 'vue'
 const atlasStore = useAtlasStore()
 const { validMapPixel } = storeToRefs(atlasStore)
 
 const props = defineProps(['lat', 'lng'])
-
-onBeforeMount(() => {
-  if (props.lat && props.lng) {
-    const latNum = parseFloat(props.lat)
-    const lngNum = parseFloat(props.lng)
-    atlasStore.setLatLngFromObject({ lat: latNum, lng: lngNum })
-    atlasStore.fetch()
-  }
-})
 </script>
